refactor base loader classes
