+++ corehq/warehouse/loaders/application.py
@@ -2,7 +2,7 @@ from corehq.apps.app_manager.models import Application
 from corehq.warehouse.const import APPLICATION_STAGING_SLUG, APPLICATION_DIM_SLUG
 from corehq.warehouse.dbaccessors import get_application_ids_by_last_modified
 from corehq.warehouse.etl import HQToWarehouseETLMixin, CustomSQLETLMixin
-from corehq.warehouse.loaders.base import BaseStagingLoader, BaseDimLoader
+from corehq.warehouse.loaders.base import BaseStagingLoader, BaseLoader
 from corehq.warehouse.models import ApplicationStagingTable, ApplicationDim
 from dimagi.utils.couch.database import iter_docs
 
@@ -39,7 +39,7 @@ class ApplicationStagingLoader(BaseStagingLoader, HQToWarehouseETLMixin):
         return iter_docs(Application.get_db(), application_ids)
 
 
-class ApplicationDimLoader(BaseDimLoader, CustomSQLETLMixin):
+class ApplicationDimLoader(BaseLoader, CustomSQLETLMixin):
     """
     Dimension for Applications
