+++ corehq/warehouse/loaders/base.py
@@ -2,11 +2,41 @@ from django.db import transaction
 
 from corehq.sql_db.routers import db_for_read_write
 from corehq.util.test_utils import unit_testing_only
-from corehq.warehouse.models.shared import WarehouseTable
 from corehq.warehouse.utils import truncate_records_for_cls
 
 
-class BaseStagingLoader(WarehouseTable):
+class BaseLoader(object):
+    model_cls = None
+
+    @classmethod
+    def commit(cls, batch):
+        """
+        Commits records based on a time frame.
+
+        :param batch: The Batch of the batch being committed
+
+        :returns: True if commit passed validations, False otherwise
+        """
+        with transaction.atomic(using=db_for_read_write(cls.model_cls)):
+            cls.load(batch)
+        return True
+
+    @classmethod
+    def dependencies(cls):
+        """Returns a list of slugs that the warehouse table is dependent on"""
+        raise NotImplementedError
+
+    @classmethod
+    def load(cls):
+        raise NotImplementedError
+
+    @classmethod
+    @unit_testing_only
+    def clear_records(cls):
+        truncate_records_for_cls(cls.model_cls, cascade=True)
+
+
+class BaseStagingLoader(BaseLoader):
     @classmethod
     def commit(cls, batch):
         cls.clear_records()
@@ -16,22 +46,3 @@ class BaseStagingLoader(WarehouseTable):
     @classmethod
     def clear_records(cls):
         truncate_records_for_cls(cls.model_cls, cascade=False)
-
-    def load(self):
-        raise NotImplementedError
-
-
-class BaseDimLoader(WarehouseTable):
-    @classmethod
-    def commit(cls, batch):
-        with transaction.atomic(using=db_for_read_write(cls.model_cls)):
-            cls.load(batch)
-        return True
-
-    @classmethod
-    @unit_testing_only
-    def clear_records(cls):
-        truncate_records_for_cls(cls.model_cls, cascade=True)
-
-    def load(self):
-        raise NotImplementedError
